Local histogram statistics replaced with integral
Local histogram extremum is not so meaningful as integral extremum. The
previous algorithm used to choose a proper orientation of semi-major axe
of inertia relied on the local maximum of the vertucal histogram of the
image rotated to yet ambiguous main axis. The orientation of the
semi-major axe was choosen consistent with the local maximum location
(below or above the center of the histogram).

The new algorithm works similar except the orientation is choosen
consisnent with the result of comparison: histogram sum below the image
center vs. histogram sum above the imega center.