Weight initialization changed
The following is applicable to XOR operation training.

rand[0, 1] weight initialization with positive sigmoid shown poor
results. The algorithm converge time was extremely large. Bipolar
sigmoid and rand[-1, 1] shown a bit better results. However the
best results were obtained with bipolar sigmoid and rand[-A, A] initial
weights, where A ~= 4..50. With A ~= 100 the algorithm did not converge
at all. With A ~= 25..50 there also is a risk the algorithm will not
converge (however there were no such cases) but with A = 50 the speed the
algorithm converges is extremely high, so the best choise is A = 50.